Introduction to Project
This project involves the development of a robotic car and a gamified feedback system that communicates with the car. The purpose of the project is aimed for students ages 8-12 to have an exciting way to engage them with feedback in numerous different ways. The feedback system is expected to provide an interactive, simple and visual way for the students to identify the data of the robotic car using a web portal. Furthermore, the data(statuses) of the car should be simple to see on the web interface in a dashboard.


Current Problems:
Lack of Technical skills
Insufficient information on the product 
Defects from the car
Integration Issue
Getting to know each member

How do we tackle such problems? 
To solve the problems, the team decided to spend at least 1 week researching and learning the features of the car and its API used for the web application. While the IS student will research the security features that can be implemented on the web application and wireless communication (wifi). The way forward for the second week is for the team to start integrating the car with the web application and setting up the web application.